fix: move type to modern/core
